server/packages/recipe_scraper/pyproject.toml
[tool.poetry]
name = "recipe-scraper"
version = "0.1.0"
description = "A command line tool to scrape and structure recipes"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "recipe_scraper", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
web-scraper = {path = "../web_scraper", develop = true}
recipe-structurer = {path = "../recipe_structurer", develop = true}
requests = "^2.32.3"
pydantic = "^2.10.0"
httpx = "^0.27.0"
argparse = "^1.4.0"
python-dotenv = "^1.0.0"
python-slugify = "^8.0.4"
datasketch = "^1.6.5"
sentence-transformers = ">=3.0.0"
numpy = ">=1.26.0"
langfuse = "^3.14.1"

[tool.poetry.scripts]
recipe-scraper = "recipe_scraper.cli:main"
test = "pytest:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.23.8" 